vt: make scrollback offset explicit

+++ include/howl_vt.h
@@ -161,7 +161,7 @@ typedef struct {
 
 int32_t howl_vt_terminal_resize(HowlVtHandle handle, uint16_t rows, uint16_t cols);
 int32_t howl_vt_terminal_ack_surface(HowlVtHandle handle, uint64_t dirty_generation);
-HowlVtSurfaceResult howl_vt_terminal_copy_surface(HowlVtHandle handle, size_t scrollback_offset, HowlVtSurfaceCell *cells_ptr, size_t cells_cap, uint8_t *dirty_rows_ptr, size_t dirty_rows_cap, uint16_t *cols_start_ptr, size_t cols_start_cap, uint16_t *cols_end_ptr, size_t cols_end_cap, uint8_t full_damage, uint16_t scroll_up_rows);
+HowlVtSurfaceResult howl_vt_terminal_copy_surface(HowlVtHandle handle, uint64_t scrollback_offset, HowlVtSurfaceCell *cells_ptr, size_t cells_cap, uint8_t *dirty_rows_ptr, size_t dirty_rows_cap, uint16_t *cols_start_ptr, size_t cols_start_cap, uint16_t *cols_end_ptr, size_t cols_end_cap, uint8_t full_damage, uint16_t scroll_up_rows);
 
 /* -------------------------------------------------------------------------- */
 /* 2. Protocol Metadata Host Output                                            */
